refactor(all): change duration to double, add logic for duration keyword

+++ src/main/kotlin/com/barbarus/prosper/core/extension/IntExtension.kt
@@ -5,4 +5,4 @@ import com.barbarus.prosper.core.activity.Duration
 object IntExtension
 
 // TODO: add .once() function
-fun Int.toDuration() = Duration(this)
+fun Double.toDuration() = Duration(this)
